fix tests for typescript
