+++ package.json
@@ -13,6 +13,7 @@
     "@types/node": "^7.0.11",
     "@types/react": "^15.0.20",
     "@types/react-dom": "^0.14.23",
+    "@types/react-test-renderer": "^15.4.2",
     "babel-cli": "^6.23.0",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
@@ -49,7 +50,7 @@
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
     "prepublish": "npm run build-lib",
-    "test-+-coverage": "react-scripts-ts test --env=jsdom --coverage --collectCoverageFrom=src/lib/**/*.{js,jsx}",
+    "test-+-coverage": "react-scripts-ts test --env=jsdom --coverage --collectCoverageFrom=src/lib/**/*.{ts,tsx}",
     "send-coverage": "cat coverage/lcov.info | codacy-coverage"
   }
 }
